changed adapter id type uint8 -> uint256

+++ src/errors/scErrors.sol
@@ -18,6 +18,6 @@ error EndUsdcBalanceTooLow();
 error FlashLoanAmountZero();
 error PriceFeedZeroAddress();
 error AmountReceivedBelowMin();
-error ProtocolNotSupported(uint8 protocolId);
-error ProtocolInUse(uint8 protocolId);
+error ProtocolNotSupported(uint256 protocolId);
+error ProtocolInUse(uint256 protocolId);
 error FloatBalanceTooLow(uint256 actual, uint256 required);
